+++ src/concrete/numpy.jl
@@ -35,11 +35,11 @@ function init_numpy()
         isreal = isint || isfloat
         isnumber = isreal || iscomplex
 
-        name = "numpy/$t"
+        name = "numpy:$t"
         rule = pyconvert_rule_numpysimplevalue{T, false}()
         saferule = pyconvert_rule_numpysimplevalue{T, true}()
 
-        pyconvert_add_rule(name, T, saferule, 100)
+        pyconvert_add_rule(name, T, saferule, PYCONVERT_PRIORITY_CANONICAL)
         isuint && pyconvert_add_rule(name, UInt, sizeof(T) ≤ sizeof(UInt) ? saferule : rule)
         isuint && pyconvert_add_rule(name, Int, sizeof(T) < sizeof(Int) ? saferule : rule)
         isint && !isuint && pyconvert_add_rule(name, Int, sizeof(T) ≤ sizeof(Int) ? saferule : rule)