+++ docs/src/releasenotes.md
@@ -1,5 +1,9 @@
 # Release Notes
 
+## Unreleased
+* `pyconvert_add_rule` is now documented. The separator of the first argument has changed
+  from `/` to `:` (not breaking because it was previously undocumented).
+
 ## v0.6.1 (2022-02-21)
 * Conversions from simple ctypes types, e.g. `ctypes.c_float` to `Cfloat`.
 * Conversions from simple numpy types, e.g. `numpy.float32` to `Float32`.