+++ docs/src/releasenotes.md
@@ -1,6 +1,6 @@
 # Release Notes
 
-## Unreleased
+## v0.7.0 (2022-03-17)
 * **Breaking:** Removes `pymethod` and `pyclass`. In the future, `pyclass` may become sugar
   for `types.new_class` (namely you can specify a metaclass).
 * Adds `pyfunc`, `pyclassmethod`, `pystaticmethod` and `pyproperty`.